Close #26 (Add tinymce5 integration)
A serious issue is that we cannot set the `file_picker_callback` via
the usual configuration.  Thus, we work around by setting up a `load`
event listener, where we destroy all TinyMCE5 editor instances, and
re-initialize them with the proper `file_picker_callback`.

The rest is similar to the TinyMCE4 integration, except that we now use
a message to communicate that the image has been picked.